+++ bin-src/tasks/initialize.ts
@@ -36,8 +36,15 @@ export const setEnvironment = async (ctx: Context) => {
 
 export const announceBuild = async (ctx: Context) => {
   const { patchBaseRef, patchHeadRef, preserveMissingSpecs } = ctx.options;
-  const { version, matchesBranch, changedFiles, replacementBuildIds, committedAt, ...commitInfo } =
-    ctx.git; // omit some fields;
+  const {
+    version,
+    matchesBranch,
+    changedFiles,
+    replacementBuildIds,
+    committedAt,
+    changedPackageManifests,
+    ...commitInfo
+  } = ctx.git; // omit some fields;
   const { rebuildForBuildId, turboSnap } = ctx;
   const autoAcceptChanges = matchesBranch(ctx.options.autoAcceptChanges);