+++ bin-src/ui/messages/errors/buildHasErrors.stories.ts
@@ -12,3 +12,22 @@ export const BuildHasErrors = () =>
     },
     exitCode: 1,
   });
+export const BuildHasErrorsAndInteractionTestFailure = () =>
+  buildHasErrors({
+    build: {
+      errorCount: 2,
+      interactionTestFailuresCount: 1,
+      webUrl: 'https://www.chromatic.com/build?appId=59c59bd0183bd100364e1d57&number=42',
+    },
+    exitCode: 1,
+  });
+
+export const BuildHasOnlyInteractionTestFailure = () =>
+  buildHasErrors({
+    build: {
+      errorCount: 2,
+      interactionTestFailuresCount: 2,
+      webUrl: 'https://www.chromatic.com/build?appId=59c59bd0183bd100364e1d57&number=42',
+    },
+    exitCode: 1,
+  });