+++ bin-src/ui/messages/errors/buildHasErrors.ts
@@ -1,15 +1,30 @@
 import chalk from 'chalk';
 import pluralize from 'pluralize';
 import { dedent } from 'ts-dedent';
-
-import { error, info } from '../../components/icons';
+import { error as errorIcon, info as infoIcon } from '../../components/icons';
 import link from '../../components/link';
 
 export default ({ build, exitCode }) => {
-  const errors = pluralize('build error', build.errorCount, true);
+  const { errorCount, interactionTestFailuresCount, webUrl } = build;
+  const hasInteractionTestFailures = interactionTestFailuresCount > 0;
+  const hasOtherErrors = errorCount - interactionTestFailuresCount > 0;
+  const failedTests = pluralize('failed test', interactionTestFailuresCount, true);
+
+  let errorMessage;
+
+  if (hasInteractionTestFailures && hasOtherErrors) {
+    const errors = pluralize('build error', errorCount - interactionTestFailuresCount, true);
+    errorMessage = `Encountered ${errors} and ${failedTests}`;
+  } else if (hasInteractionTestFailures) {
+    errorMessage = `Encountered ${failedTests}`;
+  } else {
+    const errors = pluralize('build error', errorCount, true);
+    errorMessage = `Encountered ${errors}`;
+  }
+
   return dedent(chalk`
-    ${error} {bold Encountered ${errors}}: failing with exit code ${exitCode}
+    ${errorIcon} {bold ${errorMessage}}: failing with exit code ${exitCode}
     Pass {bold --allow-console-errors} to succeed this command regardless of runtime build errors.
-    ${info} Review the errors at ${link(build.webUrl)}
+    ${infoIcon} Review the errors at ${link(webUrl)}
   `);
 };